refactor: extract shared spawn helper to deduplicate executor code
