+++ src/executor/local.ts
@@ -6,9 +6,6 @@
  * AbortController, exit code handling, and IPC broadcast.
  */
 
-import { spawn, type ChildProcess } from "child_process";
-import { createWriteStream } from "fs";
-
 import type {
   AgentExecutor,
   ExecutorSpawnOptions,
@@ -17,6 +14,7 @@ import type {
 
 import { shellSplit } from "../shell-split.ts";
 import { resolveAgentVerboseFlags } from "./agent-flags.ts";
+import { spawnChild } from "./spawn-child.ts";
 
 // ---------------------------------------------------------------------------
 // LocalExecutor
@@ -44,115 +42,30 @@ export class LocalExecutor implements AgentExecutor {
       agentVerboseFlags,
     } = opts;
 
-    return new Promise((resolve) => {
-      // Split the agent command respecting quotes
-      const parts = shellSplit(agentCommand);
-      const cmd = parts[0]!;
-      // Inject verbose flags between command parts and prompt when --verbose is active
-      const verboseFlags = verbose
-        ? resolveAgentVerboseFlags(agentCommand, agentVerboseFlags)
-        : [];
-      const args = [...parts.slice(1), ...verboseFlags, prompt];
+    // Split the agent command respecting quotes
+    const parts = shellSplit(agentCommand);
+    const cmd = parts[0]!;
+    // Inject verbose flags between command parts and prompt when --verbose is active
+    const verboseFlags = verbose
+      ? resolveAgentVerboseFlags(agentCommand, agentVerboseFlags)
+      : [];
+    const args = [...parts.slice(1), ...verboseFlags, prompt];
 
-      // Open a write stream for the agent output log (append mode).
-      // Errors are swallowed so logging never breaks the run.
-      let logStream: ReturnType<typeof createWriteStream> | undefined;
-      if (outputLogPath) {
-        try {
-          logStream = createWriteStream(outputLogPath, { flags: "a" });
-        } catch {
-          // Best-effort: if we can't open the log, continue without it
-        }
-      }
-
-      let ac: AbortController | undefined;
-      let timedOut = false;
-      const spawnOpts: {
-        cwd: string;
-        stdio: ["pipe", "pipe", "pipe"];
-        signal?: AbortSignal;
-        env?: Record<string, string | undefined>;
-      } = {
+    return spawnChild({
+      command: cmd,
+      args,
+      spawnOptions: {
         cwd,
-        stdio: ["pipe", "pipe", "pipe"],
         env: {
           ...process.env,
           HUSKY: "0",
           ...(nonce ? { RALPHAI_NONCE: nonce } : {}),
         },
-      };
-
-      if (iterationTimeout > 0) {
-        ac = new AbortController();
-        spawnOpts.signal = ac.signal;
-        setTimeout(() => {
-          timedOut = true;
-          ac!.abort();
-        }, iterationTimeout * 1000);
-      }
-
-      let child: ChildProcess;
-      try {
-        child = spawn(cmd, args, spawnOpts);
-      } catch (err) {
-        console.error(
-          `Failed to spawn agent: ${err instanceof Error ? err.message : err}`,
-        );
-        logStream?.end();
-        resolve({ output: "", exitCode: 1, timedOut: false });
-        return;
-      }
-
-      // Close stdin so the agent knows no input is coming.
-      // Without this, agents that read or wait for stdin EOF will hang.
-      child.stdin?.end();
-
-      const chunks: Buffer[] = [];
-
-      child.stdout?.on("data", (data: Buffer) => {
-        process.stdout.write(data);
-        logStream?.write(data);
-        chunks.push(data);
-        ipcBroadcast?.({
-          type: "output",
-          data: data.toString(),
-          stream: "stdout",
-        });
-      });
-
-      child.stderr?.on("data", (data: Buffer) => {
-        process.stderr.write(data);
-        logStream?.write(data);
-        chunks.push(data);
-        ipcBroadcast?.({
-          type: "output",
-          data: data.toString(),
-          stream: "stderr",
-        });
-      });
-
-      child.on("close", (code) => {
-        const output = Buffer.concat(chunks).toString("utf-8");
-        if (logStream) {
-          logStream.end(() => {
-            resolve({ output, exitCode: code ?? 1, timedOut });
-          });
-        } else {
-          resolve({ output, exitCode: code ?? 1, timedOut });
-        }
-      });
-
-      child.on("error", (err) => {
-        logStream?.end();
-        if (timedOut) {
-          const output = Buffer.concat(chunks).toString("utf-8");
-          resolve({ output, exitCode: 124, timedOut: true });
-        } else {
-          console.error(`Agent error: ${err.message}`);
-          const output = Buffer.concat(chunks).toString("utf-8");
-          resolve({ output, exitCode: 1, timedOut: false });
-        }
-      });
+      },
+      iterationTimeout,
+      outputLogPath,
+      ipcBroadcast,
+      errorLabel: "agent",
     });
   }
 }
